% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.bounds.pred.R
\name{rho.bounds.pred}
\alias{rho.bounds.pred}
\title{Estimates plausible values of the Pearson's correlation coefficient between two variables observed in distinct samples referred to the same target population.}
\usage{
rho.bounds.pred(
  data.rec,
  data.don,
  match.vars,
  y.rec,
  z.don,
  pred = "lm",
  w.rec = NULL,
  w.don = NULL,
  ...
)
}
\arguments{
\item{data.rec}{: dataframe including the Xs (predictors, listed in \code{match.vars}) and \code{y.rec} (response; target variable in this dataset).}

\item{data.don}{: dataframe including the Xs (predictors, listed in  \code{match.vars}) and \code{z.don} (response; target variable in this dataset).}

\item{match.vars}{: vector with the names of the Xs variables to be used, as (possible) predictors of respectively \code{y.rec} and \code{z.don}.}

\item{y.rec}{: character indicating the name of Y target variable in \code{data.rec}. It should be a numeric variable.}

\item{z.don}{: character indicating the name of Z target variable in \code{data.don}. It should be a numeric variable.}

\item{pred}{: method to get predictions of both Y and Z. Available methods include:
\code{pred = "lm"} (default) fits two linear regressions models (function \code{lm}) to get predictions considering respectively Y and Z as response variables and \code{match.vars} as predictors; 
\code{pred = "roblm"} (default) fits two robust linear regressions models (function \code{rlm} in package \pkg{MASS});
\code{pred = "lasso"} uses lasso method (\R package \pkg{glmnet}, function \code{\link[glmnet]{cv.glmnet}}) and cross validation to select a subset of \code{match.vars} being the best predictors of Y (Z) and then gets the models predictions;
\code{pred = "rf"} fits randomForest to get predictions of both Y and Z (function \code{\link[randomForest]{randomForest}} in \pkg{randomForest})}

\item{w.rec}{: eventual name of the variable with units' weights in \code{data.rec}; the weights are only used in estimating the correlations but not when fitting models.}

\item{w.don}{: eventual name of the variable with units' weights in \code{data.don}; the weights are only used in estimating the bounds but not when fitting models.}

\item{...}{: addition eventual parameters needed.}
}
\details{
The function assess the uncertainty in estimating the Pearson's correlation coefficient between \code{y.rec}(Y) and \code{z.don} (Z)  when the two variables are observed in two different samples referred to same target population but the samples share a number of common predictors X (\code{match.vars}). The evaluation of uncertainty corresponds to estimate the bounds (lower and upper) of the correlation coefficient between Y and Z, given the available data. The method applies the expressions suggested by Rodgers and DeVol (1982), but instead of using directly the Xs \code{match.vars}, they are replaced by predictions od both Y and Z provided by the fitted models according to \code{pred}. This latter way of working avoids drawbacks encountered in estimating covariances in presence of several X variables, where some of them are categorical (factors) and therefore pose the problem of working with the dummies. The final estimation of conditional of bounds is provided by the function \code{\link[StatMatch]{rho.bounds}}. Note that the correlations between the predictions of both Y and Z are estimated after pooling the samples. Survey weights, if available (arguments \code{w.rec} ans \code{w.don}) are used in estimating the corellations but not whend fitting the models.
}
\value{: a list with two  components: \code{corr} the estimated correlations between Y (Z) and the corresponding predicted values; \code{bounds} a vector with three values: the estimated lower bound for Pearson's correlation coefficient between \code{y.rec}(Y) and \code{z.don} (Z), the estimated upper bound and the mid-point of the interval that corresponds to the estimate Pearson's correlation coefficient under the conditional independence assumption.
}
\references{
D'Orazio, M., (2024). \emph{Is Statistical Matching feasible?} Note, \url{https://www.researchgate.net/publication/387699016_Is_statistical_matching_feasible}.

Rodgers, W.L. and DeVol E.B. (1982). An evaluation of statistical matching. \emph{Report Submitted to the Income Survey Development Program}, Dept. of Health and Human Services, Institute for Social Reasearch, University of Michigan.
}
\author{
Marcello D'Orazio \email{mdo.statmatch@gmail.com} 
}

\seealso{
  \code{\link[StatMatch]{mixed.mtc}}.
}

\examples{
set.seed(11335577)
pos <- sample(x = 1:150, size = 60, replace = FALSE)
ir.A <- iris[pos, c(1:3, 5)]
ir.B <- iris[-pos, c(1:2, 4:5)]

intersect(colnames(ir.A), colnames(ir.B)) # shared Xs

op1 <- rho.bounds.pred(data.rec=ir.A, data.don=ir.B, 
                       match.vars=c("Sepal.Length", "Sepal.Width", "Species"),
                       y.rec="Petal.Length", z.don="Petal.Width", 
                       pred = "lm")
op1
op2 <- rho.bounds.pred(data.rec=ir.A, data.don=ir.B, 
                       match.vars=c("Sepal.Length", "Sepal.Width", "Species"),
                       y.rec="Petal.Length", z.don="Petal.Width", 
                       pred = "roblm")
op2
}
