\name{StatMatch-package}
\alias{StatMatch-package}
\alias{StatMatch}
\docType{package}
\title{
Statistical Matching
}
\description{
The package StatMatch provides some R functions to perform statistical matching, i.e. the integration of two data sources referred to the same target population which share a number of common variables. Some functions can also be used to impute missing values in data sets through hot deck imputation methods. Methods to perform statistical matching when dealing with data from complex sample surveys are available too.
}
\details{
\tabular{ll}{
Package: \tab StatMatch\cr
Type: \tab Package\cr
Version: \tab 1.2.0\cr
Date: \tab 2012-12-03 \cr
License: \tab EUPL\cr
}

}
\author{
Marcello D'Orazio \cr
Maintainer: Marcello D'Orazio <madorazi@istat.it>
}
\references{
D'Orazio M., Di Zio M., Scanu M. (2006) \emph{Statistical Matching, Theory and Practice}. Wiley, Chichester.
}
